Now it is a Julia v1.0 package
