+++ src/Eff_QFI_HD.jl
@@ -1,10 +1,6 @@
 using ZChop # For chopping small imaginary parts in ρ
 using Distributed
 
-include("NoiseOperators.jl")
-include("States.jl")
-include("Fisher.jl")
-
 """
     (t, FI, QFI) = Eff_QFI_HD(Nj, Ntraj, Tfinal, dt; kwargs... )
